Add Automated IRIDA Uploads (#9)
* add initial upload work, reorganize workflows

* initial working metadata csv input

* working uploads

* add upload help statements, improve tests, add upload configs

* fixing path to test results files

* reformat tests

* actually use metadata csv in test

+++ .github/scripts/run_mpx_pipeline_metadata.sh
@@ -1,18 +1,15 @@
 #!/bin/bash
 set -eo pipefail
 
-# Create a Cache Dir
-mkdir -p conda_cache_dir
-
-# Get partial human ref genome
-gunzip $PWD/.github/ci-data/partial_hg38_ref.fa.gz
+# Test is run second so no need to re-create envs or unzip files
 
 # Run Illumina Pipeline
 nextflow run ./main.nf \
     -profile conda,test \
     --cache ./conda_cache_dir \
     --directory $PWD/.github/ci-data/ \
-    --human_ref $PWD/.github/ci-data/partial_hg38_ref.fa
+    --human_ref $PWD/.github/ci-data/partial_hg38_ref.fa \
+    --metadata_csv $PWD/.github/ci-data/test_metadata.csv
 
 # Check that the output is as expected
 # 1. Num Reads
@@ -29,9 +26,16 @@ if [[ "$COMPLETENESS" != "0.85" ]]; then
     echo "  Expected: 0.85, Got: $COMPLETENESS"
     exit 1
 fi
+# 3. Check metadata put in correctly
+CELL=`awk -F, '$1 == "Sample1" {print $7}' ./results/overall_sample_quality.csv`
+if [[ "$CELL" != "231" ]]; then 
+    echo "Incorrect output: Metadata not parsed correctly"
+    echo "  Expected: 231, Got: $CELL"
+    exit 1
+fi
 
 # Reset and Track
-mv .nextflow.log artifacts/nextflow.log
+mv .nextflow.log artifacts/nextflow2.log
 rm -rf results work/ .nextflow*
 
 echo "Passed Tests"
